LLM evaluation/requirements.txt
streamlit
groq
google-generativeai
python-dotenv
nltk
rouge-score
